requirements.txt
pandas>=2.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
streamlit>=1.27.0
plotly>=5.17.0
python-dotenv>=1.0.0
numpy>=1.24.0
loguru>=0.7.0
schedule>=1.2.0
pyyaml>=6.0.0
